# Conflicts:
#	QuanLyBanHang/src/routes.js

+++ quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dto/OrderListDto.java
@@ -6,6 +6,6 @@ import java.util.ArrayList;
 import java.util.List;
 @Data
 public class OrderListDto {
-    private Integer totalItem;
+    private Long totalItem;
     private List<OrderDto> resultItem = new ArrayList<>();
 }
